requirements.txt
python=3.8.5
numpy
pandas
matplotlib
pillow
pytorch
torchvision
torch-summary
tqdm
wandb
scikit-image
numba

black
isort
flake8
